Move all purchases code to ee directory
